+++ src/main/java/core/db/UserConfigurationTable.java
@@ -76,12 +76,9 @@ final class UserConfigurationTable extends AbstractTable {
 			insert(key, value);
 	}
 	/**
-	 * Update a key in the user configuration
-	 * if the key does not exist yet, insert it
-	 * @param key
-	 * @param value
+	 * Removes a <code>key</code> from the user configuration tablr
+	 * @param key Key to be removed.
 	 */
-	
 	void remove(String key) {
 		final StringBuffer sql = new StringBuffer(80);
 		sql.append("DELETE FROM ");
@@ -92,30 +89,6 @@ final class UserConfigurationTable extends AbstractTable {
 		sql.append(key);
 		sql.append("'");
 		adapter.executeUpdate(sql.toString());
-	}
-	
-	private String getStringValue(String key) {
-		String value = null;
-		final StringBuffer sql = new StringBuffer(100);
-		sql.append("SELECT * FROM ");
-		sql.append(getTableName());
-		sql.append(" WHERE ");
-		sql.append(columns[0].getColumnName());
-		sql.append(" = '");
-		sql.append(key);
-		sql.append("'");
-
-		final ResultSet rs = adapter.executeQuery(sql.toString());
-		try {
-			if (rs.next()) {
-				value = rs.getString(columns[1].getColumnName());
-			}
-			rs.close();
-		} catch (SQLException e) {
-			HOLogger.instance().log(getClass(), e);
-		}
-
-		return value;
 	}
 
 	private HashMap<String, String> getAllStringValues() {
@@ -190,7 +163,7 @@ final class UserConfigurationTable extends AbstractTable {
 	 * @param obj
 	 */
 	void store(Configuration obj) {
-		final HashMap<String, String> values = obj.getValues();
+		final Map<String, String> values = obj.getValues();
 		final Set<String> keys = values.keySet();
 		for (Iterator<String> iter = keys.iterator(); iter.hasNext();) {
 			String key = iter.next();
@@ -204,8 +177,8 @@ final class UserConfigurationTable extends AbstractTable {
 	 */
 	void load(Configuration obj) {
 		// initialize with default value
-		final HashMap<String,String> map = obj.getValues();
-		final HashMap<String,String> storedValues = getAllStringValues();
+		final Map<String,String> map = obj.getValues();
+		final Map<String,String> storedValues = getAllStringValues();
 
 		map.forEach((key, value) -> {
 			final String storedValue = storedValues.get(key);