+++ src/main/java/core/model/UserParameter.java
@@ -3,8 +3,8 @@ package core.model;
 import core.db.DBManager;
 import core.util.GUIUtils;
 import module.lineup.LineupAssistant;
-import java.awt.*;
 import java.util.HashMap;
+import java.util.Map;
 
 
 /**
@@ -25,16 +25,6 @@ public final class UserParameter extends Configuration {
 
     //~ Instance fields ----------------------------------------------------------------------------
 
-    //Colors for player names
-    public Color FG_STANDARD = Color.black;
-    public Color FG_TRANSFERMARKT = new Color(0, 180, 0);
-    public Color FG_INJURED = new Color(200, 0, 0);
-    public Color FG_BRUISED = new Color(100, 0, 0);
-    public Color FG_TWO_YELLOW_CARDS = new Color(100, 100, 0);
-    public Color FG_RED_CARD = new Color(200, 20, 20);
-
-    //    public String LoginName = "";
-//    public String LoginPWD = "";
     public String AccessToken = "";
     public String TokenSecret = "";
 
@@ -409,16 +399,8 @@ public final class UserParameter extends Configuration {
     }
 
     @Override
-    public HashMap<String, String> getValues() {
-        HashMap<String, String> map = new HashMap<String, String>();
-        map.put("FG_ANGESCHLAGEN", String.valueOf(FG_BRUISED.getRGB()));
-        map.put("FG_GESPERRT", String.valueOf(FG_RED_CARD.getRGB()));
-        map.put("FG_STANDARD", String.valueOf(FG_STANDARD.getRGB()));
-        map.put("FG_TRANSFERMARKT", String.valueOf(FG_TRANSFERMARKT.getRGB()));
-        map.put("FG_VERLETZT", String.valueOf(FG_INJURED.getRGB()));
-        map.put("FG_ZWEIKARTEN", String.valueOf(FG_TWO_YELLOW_CARDS.getRGB()));
-//		map.put("LoginName",String.valueOf(LoginName));
-//		map.put("LoginPWD",MyHelper.cryptString(String.valueOf(LoginPWD)));
+    public Map<String, String> getValues() {
+        Map<String, String> map = new HashMap<>();
 
         map.put("AccessToken", String.valueOf(AccessToken));
         map.put("TokenSecret", String.valueOf(TokenSecret));
@@ -636,16 +618,7 @@ public final class UserParameter extends Configuration {
     }
 
     @Override
-    public void setValues(HashMap<String, String> values) {
-        FG_BRUISED = getColorValue(values, "FG_ANGESCHLAGEN");
-        FG_RED_CARD = getColorValue(values, "FG_GESPERRT");
-        FG_STANDARD = getColorValue(values, "FG_STANDARD");
-        FG_TRANSFERMARKT = getColorValue(values, "FG_TRANSFERMARKT");
-        FG_INJURED = getColorValue(values, "FG_VERLETZT");
-        FG_TWO_YELLOW_CARDS = getColorValue(values, "FG_ZWEIKARTEN");
-
-//		LoginName = getStringValue(values,"LoginName");
-//		LoginPWD = MyHelper.decryptString(getStringValue(values,"LoginPWD"));
+    public void setValues(Map<String, String> values) {
         AccessToken = getStringValue(values, "AccessToken");
         TokenSecret = getStringValue(values, "TokenSecret");